% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_phase1.R
\name{desp1_escalation}
\alias{desp1_escalation}
\title{Accelerated escalation}
\usage{
desp1_escalation(
  data,
  cur_dose,
  ava_dose,
  lst_para,
  f_esc = tp3_escalation,
  ...
)
}
\description{
Accelerated escalation
}
