% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_c_generic.R
\name{stb_create_design}
\alias{stb_create_design}
\title{Create a Study Design}
\usage{
stb_create_design(type = c("dose_fix", "dose_3p3", "dose_acctit", "dose_boin"))
}
\description{
Create a Study Design
}
