% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb_dsgn_3p3.R
\name{tp3_escalation}
\alias{tp3_escalation}
\title{3+3 Design escalation}
\usage{
tp3_escalation(data, cur_dose, ava_dose, ...)
}
\arguments{
\item{ava_dose}{available dose levels}
}
\description{
3+3 Design escalation
}
